// ecomove-backend/src/shared/enums/transport.enums.ts
export enum TransportType {
  BICYCLE = 'bicycle',
  ELECTRIC_SCOOTER = 'electric-scooter',
  SCOOTER = 'scooter',
  ELECTRIC_VEHICLE = 'electric-vehicle'
}

export enum TransportStatus {
  AVAILABLE = 'available',
  IN_USE = 'in-use',
  MAINTENANCE = 'maintenance',
  OUT_OF_SERVICE = 'out-of-service'
}